Refactor to merge feature test into unit test
